+++ src/main/java/br/org/coletivoJava/fw/erp/implementacao/erpintegracao/json_bind_restful/SistemaERPAtual/ItfDTOSistemaERPAtual.java
@@ -55,8 +55,8 @@ public interface ItfDTOSistemaERPAtual extends ItfDTOSBJSON, ItfSistemaERPLocal 
     }
 
     @Override
-    public default int getId() {
-        return (int) getValorPorReflexao();
+    public default Long getId() {
+        return (long) getValorPorReflexao();
     }
 
     @Override
